+++ src/main/java/com/dwarfeng/meepo/util/Constants.java
@@ -30,11 +30,11 @@ public final class Constants {
     public static final String OBSERVER_EVENT_MODULE_FINISHED = "MODULE_FINISHED";
     public static final String OBSERVER_EVENT_MODULE_FAILED = "MODULE_FAILED";
 
-    public static final String POOF_CODE_SUCCESS = "0";
-    public static final String POOF_CODE_CONDITION_NOT_PASSED = "1";
-    public static final String POOF_CODE_FAILED = "2";
-    public static final String POOF_CODE_EXCEPTION = "3";
-    public static final String POOF_CODE_UNKNOWN = "4";
+    public static final int POOF_RESPONSE_CODE_SUCCESS = 0;
+    public static final int POOF_RESPONSE_CODE_EXCEPTION = 500;
+    public static final int POOF_RESPONSE_CODE_CONDITION_NOT_PASSED = 600;
+    public static final int POOF_RESPONSE_CODE_FAILED = 601;
+    public static final int POOF_RESPONSE_CODE_UNKNOWN = 900;
 
     /**
      * 换行符。